move some stuff to wb-libs
