+++ src/main/resources/reference.conf
@@ -8,7 +8,7 @@ resourceTypes =
       ownerRoleName = "admin"
       roles = {
         admin = {
-          roleActions = ["delete", "read_policies", "share_policy::admin", "share_policy::member", "share_policy::admin-notifier", "read_policy::admin", "read_policy::member", "read_policy::admin-notifier"]
+          roleActions = ["delete", "read_policies", "share_policy::admin", "share_policy::member", "share_policy::admin-notifier", "read_policy::admin", "read_policy::member", "read_policy::admin-notifier", "notify_admins"]
         },
         member = {
           roleActions = []
